+++ beacon_api/config.py
@@ -1,0 +1,8 @@
+import os
+basedir = os.path.abspath(os.path.dirname(__file__))
+
+
+class Config(object):
+    DEBUG = True
+    SQLALCHEMY_DATABASE_URI = os.environ['DATABASE_URL']
+    PUBLIC_KEY = os.environ['PUBLIC_KEY']
