+++ beacon_api/requirements.txt
@@ -1,0 +1,31 @@
+aniso8601==3.0.0
+asn1crypto==0.24.0
+atomicwrites==1.1.5
+attrs==18.1.0
+certifi==2018.4.16
+cffi==1.11.5
+chardet==3.0.4
+click==6.7
+cryptography==2.2.2
+Flask==1.0.2
+Flask-RESTful==0.3.6
+Flask-SQLAlchemy==2.3.2
+idna==2.6
+itsdangerous==0.24
+Jinja2==2.10
+jsonschema==2.6.0
+MarkupSafe==1.0
+marshmallow==2.15.2
+more-itertools==4.2.0
+pluggy==0.6.0
+py==1.5.3
+pycparser==2.18
+PyJWT==1.6.4
+pytest==3.6.0
+pytz==2018.4
+requests==2.18.4
+six==1.11.0
+SQLAlchemy==1.2.7
+urllib3==1.22
+webargs==3.0.0
+Werkzeug==0.14.1
